fixed closest_k_test_target argument error

+++ include/kmeans2d.h
@@ -42,13 +42,12 @@ void closest_n(int width, int height, int n, std::vector<float>& centers,
 * target_data_dim2 - dimension of output target points, should be same as
 *                  - target_data_dim2
 */
-void closest_k_test_target(int k, int cluster_id, float* closest,
-                           int closest_dim1, int closest_dim2, int closest_dim3,
-                           float* train_data, int train_data_dim1,
-                           int train_data_dim2, float* target_data,
-                           int target_data_dim1, int target_data_dim2,
-                           float** test, int* test_dim1, int* test_dim2,
-                           float** target, int* target_dim1, int* target_dim2);
+void closest_k_test_target(int k, int cluster_id, 
+      int* closest, int closest_dim1, int closest_dim2, int closest_dim3,
+      float* train_data, int train_data_dim1, int train_data_dim2,
+      float* target_data, int target_data_dim1, int target_data_dim2,
+      float** test, int* test_dim1, int* test_dim2,
+      float** target, int* target_dim1, int* target_dim2);
 
 void predictions_to_images(std::vector<int>& order, float* test, int test_dim1,
                            int test_dim2, float* predictions,
